+++ ui_folder/src/Components/MainContainer.scss
@@ -8,3 +8,11 @@
         margin: 2rem 0;
     }
 }
+
+#loading-receipt-spinner {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    min-height: 5rem;
+}
